Reduce shortcut client payload and tune Lighthouse
