+++ .lighthouserc.json
@@ -21,7 +21,7 @@
         "categories:performance": [
           "error",
           {
-            "minScore": 0.9
+            "minScore": 0.85
           }
         ],
         "categories:accessibility": [
